rpi: Support both UART interfaces on Raspberry Pi 3.
The BCM2837 has two serial port interfaces, one supported by the PL01X
driver (the serial0) and one supported by BCM283X_MU driver (the
serial1 or mini-uart). The Raspberry Pi 3 default's device tree
connects the serial0 to the Bluetooth module and leaves the mini-uart
to the external GPIO port. The mini-uart has problems since its
frequency depends on the CPU frequency.

Via a device tree overlay, it is possible to change the serial0 back to
the GPIO pins if for example Bluetooth is not needed. This patch adds
support for both serial ports in the Raspberry Pi 3 and decides at
runtime which one to enable based on the value of the RX pin muxing in
the GPIO port.

Test: Booted a rpi3 with "enable_uart=1" and with
"dtoverlay=pi3-miniuart-bt" in the config.txt. Both output serial data.

Change-Id: I84bfd48ff837be548487f931572181ee5868da86