Add Android.bp to build U-Boot's mkimage.
This tool is needed to build FIT images for the Beagleboard-X15
platform. As it is not a required host build dependency, we must build
it from sources.

This change introduces a build of the mkimage tool from 2018.07, however
the image signing feature is disabled because it depends on an openssl
API which was removed from boringssl. This would need further changes
upstream to resolve.

A directory containing a large number of symlinks to existing headers is
also provided, as the GNU Make build system was using the "-idirafter"
feature to stop copies of standard C headers like stdint.h, stdlib.h
from being selected from the U-Boot 'include' directory, which is an
unsupported feature in the Android build system.

Bug: 74437007
Change-Id: Idadda650232749d198320b4556a7ba5379a441d0
Signed-off-by: Alistair Strachan <astrachan@google.com>